Functionality to fetch rendered file attachment
Added functionality to render file attachment and return the rendered
HTML as plain-text so it can be dynamically added to the DOM on the
frontend.

Each Review UI can now set a 'urlpatterns' property. Each URL pattern
that is set for any review UI gets appended to the URL patterns in
reviews/urls.py and hence gets handled by Django when the URL is
accessed.